refactor: 패키지 구조 변경
service 패키지 추가.
