+++ src/main/java/lotto/domain/LottoNumberPool.java
@@ -1,4 +1,4 @@
-package lotto.domain.core;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
