+++ src/main/java/lotto/domain/Ticket.java
@@ -1,0 +1,8 @@
+package lotto.domain;
+
+public interface Ticket {
+
+    LottoNumbers ticketNumbers();
+
+    boolean contains(LottoNumber number);
+}
