--- src/main/java/lotto/domain/core/Ticket.java
@@ -1,8 +1,0 @@
-package lotto.domain.core;
-
-public interface Ticket {
-
-    public TicketNumbers ticketNumbers();
-
-    public boolean contains(TicketNumber number);
-}
