--- src/main/java/lotto/domain/core/TicketNumber.java
@@ -1,6 +1,0 @@
-package lotto.domain.core;
-
-public interface TicketNumber extends Comparable<TicketNumber> {
-
-    int get();
-}
