--- src/main/java/lotto/domain/core/TicketNumbers.java
@@ -1,10 +1,0 @@
-package lotto.domain.core;
-
-import java.util.List;
-
-public interface TicketNumbers {
-
-    boolean contains(TicketNumber number);
-
-    List<TicketNumber> numbers();
-}
