--- src/main/java/lotto/domain/core/UserTickets.java
@@ -1,11 +1,0 @@
-package lotto.domain.core;
-
-import lotto.domain.LottoResult;
-
-import java.util.List;
-
-public interface UserTickets {
-    List<Ticket> tickets();
-
-    LottoResult result(WinningTicket winningTicket);
-}
