--- src/main/java/lotto/domain/core/WinningTicket.java
@@ -1,8 +1,0 @@
-package lotto.domain.core;
-
-public interface WinningTicket {
-
-    public int match(Ticket ticket);
-
-    public boolean bonus(Ticket ticket);
-}
