+++ src/main/java/lotto/service/UserLottosCreator.java
@@ -1,6 +1,6 @@
-package lotto.domain;
+package lotto.service;
 
-import lotto.domain.core.*;
+import lotto.domain.*;
 import lotto.domain.exceptions.ExceptionMessages;
 import lotto.domain.exceptions.LottoTicketException;
