+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
-import lotto.domain.core.Rank;
 import lotto.dto.LottoResultDto;
+import lotto.service.LottoResult;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
