+++ src/test/java/lotto/domain/RankTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.core.Rank;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
