+++ src/test/java/lotto/domain/WinningLottoTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.core.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
