+++ src/test/java/lotto/domain/core/LottoNumberTest.java
@@ -1,5 +1,6 @@
 package lotto.domain.core;
 
+import lotto.domain.LottoNumber;
 import lotto.domain.exceptions.LottoNumberException;
 import org.junit.jupiter.api.Test;
